+++ app/Http/Controllers/Customer/ListingController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers\Customer;
 
 use Inertia\Inertia;
-use App\Models\Space;
+use App\Queries\SpaceQuery;
 use App\Filters\SpaceFilter;
 use App\Http\Controllers\Controller;
 
@@ -12,14 +12,15 @@ class ListingController extends Controller
     /**
      * Display a listing of the resource.
      *
+     * @param \App\Queries\SpaceQuery  $query
      * @param \App\Filters\SpaceFilter $filters
      *
      * @return \Illuminate\Http\Response
      */
-    public function __invoke(SpaceFilter $filters)
+    public function __invoke(SpaceQuery $query, SpaceFilter $filters)
     {
         return Inertia::render('Welcome/Index', [
-            'products' => Space::listing($filters)->paginate(),
+            'products' => $query->listing($filters)->paginate(),
         ]);
     }
 }
