+++ app/Models/Space.php
@@ -2,15 +2,12 @@
 
 namespace App\Models;
 
-use App\Filters\SpaceFilter;
 use App\Models\Traits\Orderable;
 use App\Models\Casts\ScheduleCast;
 use App\Models\Traits\Productable;
-use Illuminate\Support\Facades\DB;
 use App\Models\Concerns\ManagesProduct;
 use Illuminate\Database\Eloquent\Model;
 use App\Models\Traits\HasEncryptableCode;
-use Illuminate\Database\Eloquent\Builder;
 use Cratespace\Preflight\Models\Traits\Directable;
 use Cratespace\Preflight\Models\Traits\Filterable;
 use Cratespace\Preflight\Models\Traits\Presentable;
@@ -90,31 +87,5 @@ class Space extends Model implements ProductContract
     public function owner(): BelongsTo
     {
         return $this->belongsTo(User::class, 'user_id');
-    }
-
-    /**
-     * Get a listing of all available spaces.
-     *
-     * @param \App\Filters\SpaceFilter $filters
-     *
-     * @return \Illuminate\Database\Eloquent\Builder
-     */
-    public static function listing(SpaceFilter $filters): Builder
-    {
-        return static::query()->addSelect([
-                'business' => Business::select('name')
-                    ->whereColumn('user_id', 'spaces.user_id')
-                    ->latest()
-                    ->take(1),
-            ])
-            ->whereNotExists(function ($query) {
-                $query->select(DB::raw(1))
-                    ->from('orders')
-                    ->whereColumn('orders.orderable_id', 'spaces.id');
-            })
-            ->whereNull('reserved_at')
-            ->whereDate('departs_at', '>', now())
-            ->filter($filters)
-            ->latest('departs_at');
     }
 }
