+++ app/Queries/SpaceQuery.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace App\Queries;
+
+use App\Models\Space;
+use App\Models\Business;
+use App\Filters\SpaceFilter;
+use Illuminate\Support\Facades\DB;
+use Cratespace\Preflight\Queries\Query;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Builder;
+
+class SpaceQuery extends Query
+{
+    /**
+     * Instance of model being queried.
+     *
+     * @var \Illuminate\Database\Eloquent\Model
+     */
+    protected $model = Space::class;
+
+    /**
+     * Get a listing of all available spaces.
+     *
+     * @param \App\Filters\SpaceFilter $filters
+     *
+     * @return \Illuminate\Database\Eloquent\Builder
+     */
+    public function listing(SpaceFilter $filters): Builder
+    {
+        return $this->model()
+            ->query()
+            ->addSelect([
+                'business' => Business::select('name')
+                    ->whereColumn('user_id', 'spaces.user_id')
+                    ->latest()
+                    ->take(1),
+            ])
+            ->whereNotExists(function ($query) {
+                $query->select(DB::raw(1))
+                    ->from('orders')
+                    ->whereColumn('orders.orderable_id', 'spaces.id');
+            })
+            ->whereNull('reserved_at')
+            ->whereDate('departs_at', '>', now())
+            ->filter($filters)
+            ->latest('departs_at');
+    }
+}
