+++ phpunit.xml
@@ -2,8 +2,7 @@
 <phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:noNamespaceSchemaLocation="./vendor/phpunit/phpunit/phpunit.xsd"
          bootstrap="vendor/autoload.php"
-         colors="true"
->
+         colors="true">
     <testsuites>
         <testsuite name="Unit">
             <directory suffix="Test.php">./tests/Unit</directory>
@@ -13,6 +12,9 @@
         </testsuite>
         <testsuite name="Console">
             <directory suffix="Test.php">./tests/Console</directory>
+        </testsuite>
+        <testsuite name="Inertia">
+            <directory suffix="Test.php">./tests/Inertia</directory>
         </testsuite>
     </testsuites>
     <coverage processUncoveredFiles="true">
