+++ tests/Unit/Queries/OrderQueryTest.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Tests\Unit\Queries;
+
+use PHPUnit\Framework\TestCase;
+
+class OrderQueryTest extends TestCase
+{
+    /**
+     * A basic unit test example.
+     *
+     * @return void
+     */
+    public function test_example()
+    {
+        $this->assertTrue(true);
+    }
+}
